// ui/src/api/generated/models/test-media-content.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.26.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TestMediaContent
 */
export interface TestMediaContent {
    /**
     * 
     * @type {string}
     * @memberof TestMediaContent
     */
    'data'?: string;
    /**
     * 
     * @type {string}
     * @memberof TestMediaContent
     */
    'filename'?: string;
    /**
     * 
     * @type {string}
     * @memberof TestMediaContent
     */
    'mediaType'?: string;
    /**
     * 
     * @type {string}
     * @memberof TestMediaContent
     */
    'url'?: string;
}